http_prebuilt: Improve retrieval of artifacts from archive
The current version always redownloads the artifact.
Kali does not like the nested md5sum command and ignores it
leaving file checksum value blank so the test always fails.
A temporary variable is used to get around this limitation
of kali.

The echo "Downloading..." command has been modified to actually
print and has been moved so it only does so when downloading
the artifact.

Change-Id: I115860a87e55399a652d6548fe50f6204337f854

# Conflicts:
#	build/core/http_prebuilt.mk